# Making performance improvements to sockets