@davidfowl 2018-03-06
Failures:

- OSX: Broken with the known issue
- on Windows we have a flaky libuv test:

```
dotnet.exe : [xUnit.net 00:00:09.1627717]     WriteAsyncWithTokenAfterCallWithoutIsCancelled(maxResponseBufferSize: 1024) [FAIL]
325At C:\Users\appveyor\.dotnet\buildtools\korebuild\2.1.0-preview2-15726\scripts\common.psm1:7 char:5
326+     & $_cmd @args
327+     ~~~~~~~~~~~~~
328    + CategoryInfo          : NotSpecified: ([xUnit.net 00:0...e: 1024) [FAIL]:String) [], RemoteException
329    + FullyQualifiedErrorId : NativeCommandError
330 
331[xUnit.net 00:00:09.1632381]       Assert.True() Failure
332[xUnit.net 00:00:09.1633962]       Expected: True
333[xUnit.net 00:00:09.1635380]       Actual:   False
334[xUnit.net 00:00:09.1645863]       Stack Trace:
335[xUnit.net 00:00:09.1658683]         C:\projects\kestrelhttpserver\test\Kestrel.Transport.Libuv.Tests\LibuvOutputConsumerTests.cs(562,0): at Microsoft.AspNetCore.Server.Kestrel.Transport.Libuv.Tests.LibuvOutputConsumerTests.<>c__DisplayClass16_0.<<WriteAsyncWithTokenAfterCallWithoutIsCancelled>b__0>d.MoveNext()
336[xUnit.net 00:00:09.1660616]         --- End of stack trace from previous location where exception was thrown ---
337[xUnit.net 00:00:09.1662406]         C:\projects\kestrelhttpserver\test\Kestrel.Transport.Libuv.Tests\LibuvOutputConsumerTests.cs(505,0): at Microsoft.AspNetCore.Server.Kestrel.Transport.Libuv.Tests.LibuvOutputConsumerTests.WriteAsyncWithTokenAfterCallWithoutIsCancelled(Int32 maxResponseBufferSize)
338[xUnit.net 00:00:09.1664034]         --- End of stack trace from previous location where exception was thrown ---
339Failed   WriteAsyncWithTokenAfterCallWithoutIsCancelled(maxResponseBufferSize: 1024)
340Error Message:
341 Assert.True() Failure
342Expected: True
343Actual:   False
344Stack Trace:
345   at Microsoft.AspNetCore.Server.Kestrel.Transport.Libuv.Tests.LibuvOutputConsumerTests.<>c__DisplayClass16_0.<<WriteAsyncWithTokenAfterCallWithoutIsCancelled>b__0>d.MoveNext() in C:\projects\kestrelhttpserver\test\Kestrel.Transport.Libuv.Tests\LibuvOutputConsumerTests.cs:line 562
346--- End of stack trace from previous location where exception was thrown ---
347   at Microsoft.AspNetCore.Server.Kestrel.Transport.Libuv.Tests.LibuvOutputConsumerTests.WriteAsyncWithTokenAfterCallWithoutIsCancelled(Int32 maxResponseBufferSize) in C:\projects\kestrelhttpserver\test\Kestrel.Transport.Libuv.Tests\LibuvOutputConsumerTests.cs:line 505
348--- End of stack trace from previous location where exception was thrown ---
349[xUnit.net 00:00:11.4722859]   Finished:    Microsoft.AspNetCore.Server.Kestrel.Transport.Libuv.Tests
350[xUnit.net 00:00:00.0049864] xUnit.net VSTest Adapter v2.4.0-beta.1.build3945 (64-bit .NET Core 4.6.26222.07)
351
```

I believe this is one that @pakrym encountered when porting to the latest pipelines

- Linux - Test run aborted without a message, this is pretty scary and I've seen it a few times now.