@halter73 2018-03-07
So I talked to @davidfowl about the purpose of double dispatching, and his theory is that without this change, a single socket could dominate the usage of the limited I/O threads for as long as there's data already available in the OS's input buffer by the time ReceiveAsync is called.

I figured if that's the case, certain connections must be starving others of resources in dev, and that would mean this change should reduce wrk's reported standard deviation for latency. It looks like @davidfowl is on to something, because the latency numbers for the Socket transport are immensely better with this PR.

 
## Json Windows

### Sockets

#### Before

```
[01:17:11.217] Starting 'Wrk' worker
[01:17:11.217] Measuring first request latency on http://10.0.0.110:5000/json
[01:17:11.219] 1.5123 ms
[01:17:11.219] Measuring subsequent requests latency
[01:17:11.227] 0.7173 ms
[01:17:11.227] wrk -H "Host: localhost" -H "Accept: application/json,text/html;q=0.9,application/xhtml+xml;q=0.9,application/xml;q=0.8,*/*;q=0.7" -H "Connection: keep-alive" --latency -d 15 -c 256 --timeout 8 -t 32  http://10.0.0.110:5000/json
[01:17:11.290] Running 15s test @ http://10.0.0.110:5000/json
[01:17:11.290]   32 threads and 256 connections
[01:17:26.398]   Thread Stats   Avg      Stdev     Max   +/- Stdev
[01:17:26.408]     Latency     7.18ms   93.79ms   2.64s    99.32%
[01:17:26.408]     Req/Sec     9.29k   503.29    14.79k    85.89%
[01:17:26.408]   Latency Distribution
[01:17:26.408]      50%    0.85ms
[01:17:26.408]      75%    0.87ms
[01:17:26.408]      90%    0.90ms
[01:17:26.408]      99%   10.06ms
[01:17:26.408]   4439914 requests in 15.10s, 643.60MB read
[01:17:26.408] Requests/sec: 294054.23
[01:17:26.408] Transfer/sec:     42.63MB
[01:17:26.518] Failed to parse bad responses
[01:17:26.518] Failed to parse bad responses
[01:17:27.113] Deleting job [ID:258 Connections:256 Threads:32 Duration:15 Method:GET ServerUrl:http://10.0.0.110:5000/json Headers:{"Host":"localhost","Accept":"application/json,text/html;q=0.9,application/xhtml+xml;q=0.9,application/xml;q=0.8,*/*;q=0.7","Connection":"keep-alive"}]
```

#### After

```
[01:18:44.580] Starting 'Wrk' worker
[01:18:44.580] Measuring first request latency on http://10.0.0.110:5000/json
[01:18:44.581] 1.0201 ms
[01:18:44.581] Measuring subsequent requests latency
[01:18:44.587] 0.6895 ms
[01:18:44.587] wrk -H "Host: localhost" -H "Accept: application/json,text/html;q=0.9,application/xhtml+xml;q=0.9,application/xml;q=0.8,*/*;q=0.7" -H "Connection: keep-alive" --latency -d 15 -c 256 --timeout 8 -t 32  http://10.0.0.110:5000/json
[01:18:44.650] Running 15s test @ http://10.0.0.110:5000/json
[01:18:44.650]   32 threads and 256 connections
[01:18:59.752]   Thread Stats   Avg      Stdev     Max   +/- Stdev
[01:18:59.755]     Latency     1.36ms   14.06ms 865.28ms   99.56%
[01:18:59.755]     Req/Sec    11.07k   621.43    15.27k    89.95%
[01:18:59.755]   Latency Distribution
[01:18:59.755]      50%  706.00us
[01:18:59.755]      75%  741.00us
[01:18:59.755]      90%  786.00us
[01:18:59.756]      99%    0.99ms
[01:18:59.756]   5265427 requests in 15.10s, 763.27MB read
[01:18:59.756] Requests/sec: 348722.18
[01:18:59.756] Transfer/sec:     50.55MB
[01:18:59.887] Failed to parse bad responses
[01:18:59.887] Failed to parse bad responses
[01:19:00.452] Deleting job [ID:260 Connections:256 Threads:32 Duration:15 Method:GET ServerUrl:http://10.0.0.110:5000/json Headers:{"Host":"localhost","Accept":"application/json,text/html;q=0.9,application/xhtml+xml;q=0.9,application/xml;q=0.8,*/*;q=0.7","Connection":"keep-alive"}]
```

So not only did this chage reduce the standard deviation for latency from 93.79ms to 14.06ms, it also reduced the average latency from 7.18ms to 1.36ms.

The 99th latency percentile is even more impressive. The 50/75/90th percentiles are slightly better than before (note the difference between ms and us), but the 99th percentile went from 10.06ms to 0.99ms! It's interesting that the 99th percentile for latency is still lower than the average after than this change. It really shows that my stats prof was right about the mean being heavily impacted by outliers.